Add server shutdown support
Until now it was impossible to shutdown the HTTP server once started (due
to the use of the implicitly created server in `http.ListenAndServe()`).

This changes things to explicitly create and use dedicated `http.Server`
instances for both the regular `Server` and `MuxServe` versions.

This allows to later call `.Shutdown()` for a clean shutdown.